Item Position Simulation in a Preference Test

With simsim function the putative position of new items in a preference test can be simulated based on its Worth Value. This requires a minimum set of tested items and the remaining combinations with the ground truth (GT) will be simulated. The performance of the simulation can be assessed with the amount of intransitive results (expressed as a ratio: Iratio) and the overall consensus error (CE). The more experimentally tested item combinations enter the simulation, the more accurate the simulated position will be.

simsim(
  data = NULL,
  simOpt = NULL,
  GT = NULL,
  seeding = TRUE,
  runs = NULL,
  truepos = NULL,
  verbose = FALSE,
  path = NULL
)

Arguments

data
rawdata from the bimload function

simOpt
item to be checked (this can be an item from the GT or a new one. Remove the item from GT if specified here!)

GT
item list with the ground truth (GT; letters are case sensitive!)

seeding
use constant seeding for stable simulation results (default=TRUE)

runs
the number of simulation runs (e.g., 200)

truepos
the true position of the item (most likely unknown); used for performance testing

verbose
FALSE silences the bimpre output on the number of simulated ties

path
path where the results of the simulations are saved as *.txt files. If left empty (NULL), the results will only be stored in memory.

filter.crit
the filter criterion for the plot (defaults to "CE": consensus error; alternatively the intransitivity ratio (Iratio) can be shown)

Value

list with the position frequencies and the plot object